INF-3389 Use query to find upgraded APs
Adds query to get the earliest QDate per AP where a job reported
CondorVersion > 24.11.1 to figure out when to expect first job
ads containing NumVacatesByReason. Then use the results of the
query when deciding if jobs should be used in the calcualtion
of number of jobs interrupted due to input transfer failure.