INF-3411 Distinguish director (or plugin) errors
Now that the ErrorType is being recorded for transfer attempts
that don't make it past the director contact stage, we can
distinguish pre-endpoint contact errors... sometimes. We're still
both waiting for the version with the ErrorType definitions to be
rolled out *and* also dealing with the ErrorType being "Unprocessed"
(essentially undefined) in some cases.